+++ NEWS.md
@@ -1,3 +1,16 @@
+# crew.aws.batch 0.0.7
+
+* Send both cancellation and termination requests to end jobs.
+* Fix launcher bug/typo where parameters were supplied to container overrides.
+* Add a new `all` argument to `terminate()` in the AWS Batch monitor.
+* Add `r_arguments` to accept command line arguments to R.
+* Support `options_metrics`.
+* Reduce argument clutter with `crew_options_aws_batch()`. Supports direct inputs for CPUs, GPUs, and memory without having to specify a complicated `containerOverrides` list.
+* Sanitize job names.
+* Use `crashes_error` from `crew`.
+* Make `cpus`, `gpus`, and `memory` retryable options.
+* Change default `seconds_idle` to 300.
+
 # crew.aws.batch 0.0.6
 
 * Add a `retry_tasks` argument.